refactor: Form과 관련 없는 요소들 분리
