+++ frontend/src/pages/GuestReservation/GuestReservation.tsx
@@ -2,8 +2,13 @@ import { useEffect } from 'react';
 import { useHistory, useLocation, useParams } from 'react-router-dom';
 import Header from 'components/Header/Header';
 import Layout from 'components/Layout/Layout';
+import PageHeader from 'components/PageHeader/PageHeader';
+import ReservationListItem from 'components/ReservationListItem/ReservationListItem';
+import MESSAGE from 'constants/message';
+import useGuestReservations from 'hooks/useGuestReservations';
 import { GuestMapState } from 'pages/GuestMap/GuestMap';
 import { MapItem, Reservation, ScrollPosition, Space } from 'types/common';
+import * as Styled from './GuestReservation.styles';
 import ReservationForm from './ReservationForm';
 
 interface GuestReservationState {
@@ -27,6 +32,9 @@ const GuestReservation = (): JSX.Element => {
 
   if (!mapId || !space) history.replace(`/guest/${sharingMapId}`);
 
+  const getReservations = useGuestReservations({ mapId, spaceId: space.id, date: selectedDate });
+  const reservations = getReservations.data?.data?.reservations ?? [];
+
   useEffect(() => {
     return history.listen((location) => {
       if (
@@ -46,6 +54,10 @@ const GuestReservation = (): JSX.Element => {
     <>
       <Header />
       <Layout>
+        <Styled.PageHeader title="공간 이름" data-testid="spaceName">
+          <Styled.ColorDot color={space.color} />
+          {space.name}
+        </Styled.PageHeader>
         <ReservationForm
           mapId={mapId}
           space={space}
@@ -53,6 +65,25 @@ const GuestReservation = (): JSX.Element => {
           sharingMapId={sharingMapId}
           reservation={reservation}
         />
+        <Styled.Section>
+          <PageHeader title={`${selectedDate}${selectedDate && '의'} 예약 목록`} />
+          {getReservations.isLoadingError && (
+            <Styled.Message>{MESSAGE.RESERVATION.ERROR}</Styled.Message>
+          )}
+          {getReservations.isLoading && !getReservations.isLoadingError && (
+            <Styled.Message>{MESSAGE.RESERVATION.PENDING}</Styled.Message>
+          )}
+          {getReservations.isSuccess && reservations.length === 0 && (
+            <Styled.Message>{MESSAGE.RESERVATION.SUGGESTION}</Styled.Message>
+          )}
+          {getReservations.isSuccess && reservations.length > 0 && (
+            <Styled.ReservationList role="list">
+              {reservations?.map((reservation) => (
+                <ReservationListItem key={reservation.id} reservation={reservation} />
+              ))}
+            </Styled.ReservationList>
+          )}
+        </Styled.Section>
       </Layout>
     </>
   );
